pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fermi"
version = "0.1.0"
description = "Automated diagram-based question generation pipeline for Grades 9-12"
requires-python = ">=3.9"
dependencies = [
    "openai==1.3.0",
    "pydantic==2.5.0",
    "pypdfium2==4.17.0",
    "Pillow==10.1.0",
    "PyYAML==6.0.1",
    "reportlab==4.0.7",
    "streamlit>=1.28.0",
    "groq>=0.4.1",
    "python-dotenv>=1.0.0",
]

[project.scripts]
generate = "src.cli:main"
